Fix XER to JSON date parser throwing wrong type of exception
The XER to JSON date parser would throw a "Bad type" exception if one of
the fields is missing, but would not check if the actual date node was
present. Thus, the optional modifier was completely useless.

Now, the parser will first check if the node is present, and if not,
throw a "not found" exception. If the node is found, but a field is
missing, it will still throw a "bad type" exception as expected.